layout/style/test/test_bug573255.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=573255
-->
<head>
  <title>Test for Bug 573255</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">

  #display:not(.hasSummary) { visibility: hidden }

  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=573255">Mozilla Bug 573255</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

var p = document.getElementById("display");
var cs = getComputedStyle(p, "");

is(cs.visibility, "hidden", "should be visibility:none since it has no class");
p.className = "hasSummary";
is(cs.visibility, "visible", "changing class attribute should remove visibility:hidden");

</script>
</pre>
</body>
</html>
